+++ lib/pg_triggers.rb
@@ -20,6 +20,9 @@ module PgTriggers
   mattr_accessor :permission_checker
   self.permission_checker = nil
 
+  mattr_accessor :excluded_tables
+  self.excluded_tables = []
+
   def self.configure
     yield self
   end
@@ -28,10 +31,11 @@ module PgTriggers
   autoload :DSL, "pg_triggers/dsl"
   autoload :Registry, "pg_triggers/registry"
   autoload :Drift, "pg_triggers/drift"
-  autoload :Audit, "pg_triggers/audit"
   autoload :Permissions, "pg_triggers/permissions"
   autoload :SQL, "pg_triggers/sql"
   autoload :DatabaseIntrospection, "pg_triggers/database_introspection"
   autoload :Generator, "pg_triggers/generator"
   autoload :Testing, "pg_triggers/testing"
+  autoload :Migration, "pg_triggers/migration"
+  autoload :Migrator, "pg_triggers/migrator"
 end